Add mock.rs
